Improve ruby:bob test suite and example code
The new example has two teenagers: Alice and Bob.

They both have the same functionality. Bob is implemented in the most
straight-forward way possible with an if/else, whereas Alice is
implemented with separate objects to represent each type of answer, and
that response also knows whether or not it can handle a certain type of
input.

This would be a good solution if the code were expected to change or be
extended.