Refactor emails
Since emails are very similar to each other it seems like it would make
sense to use inheritance. The dispatch class was adding a bit of
indirection, and had to pass to much hand-crafted information to the
dispatch instance, which then passed it on to the email instance.

I think this might end up being a bit cleaner in the long run. We'll
see.