Deprecate old summary url
The summary page for an assignment was /user/:language/:slug, which means
that only the current user could access a particular summary.

By changing this to be /:username/:language/:slug we can allow
nitpickers to view the full history for an assignment for another user.

Redirect from the old url to the new one so that people can still click
links from old emails.