Discourage empty, feel-good nitpicks
"Good job" or "great work" are not helpful feedback. If it's used to
just soften the following criticisms, that is going to be obvious to the
reader.

Be specific about what you like about a solution. Simplicity? The
expressive clarity of a variable name? The choice of an enumerable
method?