Show count of nits by others and revision number
This suggests changes to the data displayed on the dashboard. Now:

- Nits are counted only if they are submitted by others, so a user can
  comment on their own code without "penalty".

- The arguments/discussions count only goes up when the user
  participates in the discussion. This is a different kind of signal
  than a third user chiming in, as it shows that the user has
  (implicitly) acknowledged the feedback of the nit.

- A new number is shown reflecting the revision count of the submission.
  This helps a potential nitpicker better distinguish early code from
  code that is likely closer to the end.

Basic tests are added for each kind of data. The routines that sort
pending requests have also been adjusted to conform to this suggestion.

(I cannot figure out the best way to test the views on a local
installation. I'm sorry about that. I would be happy to work through
that if someone can help me figure out how to access a local
installation as an admin user, preferably without creating too many fake
github accounts.)

(I also do not know how this performs with MongoDB. I followed
the style of other methods and used simple selects or maps based on the
retrieved data.)