Add test for same length, same letters, different letter instance counts
One possible solution to detecting an anagram might be to check to see if the words are the same length and that each letter from the first word is contained in the second word. This additional test would guard against that solution appearing to work.

I wasn't sure in what position in the test sequence it ought to go, or what the exact test description should be. Feel free to suggest changes.