Update how nits are counted for a user's own work
Right now, the nit count shown includes a user's own nits.  A user may
see this and assume that the nits they add as commentary are having the
effect of pushing their own assignments off the Featured page, because
they see the "1" next to their own work and the column of "0" on the
Featured page.

This commit corrects that. A user's own nits now do not increase the
count shown in the sidebar.