Always nitpicking with direct link. Fixes #507.
If you have a direct link to a submission, you will be allowed to
view and comment (and click "looks great"), provided that you are logged
in.

In order to make this a reasonable feature, the access was removed from
the /:user/:language/:exercise page unless you have completed the
exercise, and any exercises that you've not completed will not have
links in the user's overview page.