Fix scrollbars appear while opening notifications
When opening the notification sidebar in some browsers and OS there will be horizontal scrollbars appearing for the duration of the opening animation. This is cause by the notification list dangling ooutside of the window. By setting the body's overflow-x to hidden when the panel is out, these scrollbars are gone. This problem fixed here was probably introduced with 0215134 when the animation was enabled.