Prefer recent exercises in random work
It's more likely that there are fewer people nitpicking on the recent
exercises.

This does not loop through multiple languages, so there's a chance that
you will end up with no suggestion. This will be fixed in a later
commit.

Also, this is a hack. There are no tests. This will also be fixed in
said later commit.