Add test for numbers with greater than 11 digits that start with 1
I came across an implementation where the validation was written in such
a way that any numbers greater than 10 digits in length that started with
1 were considered valid. Without this extra test extremely long numbers
that happen to start with 1 could all be considered valid and could be
passed through with just the 1 removed from the front. (That is how the
implementation I came across was doing it: `number[0] = ""`).