Clojure ETL full-dataset test incorrectly removes first element
The form of the test data uses a subvec with an index of 1:
(subvec (clojure.string/split "AEIOULNRST" #"") 1)
That removes the first element from the list so "A" is lost and
the current example fails that test. Fix it so the initial letter
is not removed and example is correct for test.

Signed-off-by: Geoff Lane <geoff@zorched.net>