Apply clearfix so the first nit's first line doesn't wrap
The form element inside the h2 element has 20px bottom margin and it's
floating, which in this case means its expanding outside of it's
container and causing the text from the first nit to wrap around it.
This makes the first line of the first nit wrap about half way across
the line.

Just applied bootstrap's clearfix class to the h2 to make the h2 expand
to contain the form. I could have adjusted the bottom margin on the
form, too, but it seems ok, especially with 1.0 coming soon.