Do not redirect to new exercises API
It appears that I've configured something wrong, so that a redirect to a
different subdomain fails in the CLI. It works whon both are
`localhost`, and I assumed that it would work when `exercism.io` is
delegating to `x.exercism.io`.

Until I can figure this out, I am making the actual request from the
controller and then serving the response from the new API as the
response from the controller action.