Make submissions visible to team managers.
When a team manager is looking at the team overview page, each
exercise should link directly to the submission page for that
exercise.

On a user's profile page, all the exercises should be properly linked if
the person looking at it manages a team that the user is a member of.