Convert all Ruby hashes to 1.9 sexy syntax
All the hashes with key-value pairs in the form of:

  :symbol_key => 'a value'

have been converted to the "sexy syntax" introduced in Ruby since version 1.9,
that is:

  symbol_key: 'a value'

I think it's cleaner and it makes code more readable while not affecting it in
any other way. It's not a performance effort in any way.