Remove duplication in use of scopes.
The only difference between the `pending` and the `needs_input` method
is the use of the scope for what state the submission is in. Since we're
dealing with unmuted submissions for a particular language, give that
concept a name via a method and use it to remove the common scopes used
in both the `pending` and `needs_input` methods.