Retina #brand logo in the navbar
I replaced the `e_small.png` and `e_red_small.png` images with the same images
but with their sizes doubled. This way the logo looks nice on retina screens. I
also had to modify some CSS in order to constrain the images (which are just
`background`s) into a fixed size, since the previous CSS relied only on the
actual size of the images.