Allow unauthenticated guests to view a submission
Allow unauthenticated guests to view all the iterations of code and all
the nitpicks others have left on the submission page.

Allow unauthenticated guests to see a lot of links on the submission
page, to give them an idea of what they could do if they signed up.
Most of these links lead to a "you have to be logged in to do that"
page.

Show nitpick page title and nav links even if no submissions are present
in list.  This improves the UX of that page a little but most
importantly allows us to make `NullWorkload` work with unauthenticated
guests.