Return the iteration number when submitting an exercise.
Currently we always tell people to fetch again, whether or not a new
problem has been unlocked. See #2315.

If this is the first iteration, then we want to tell the user that they
can fetch the next exercise, but if it's resubmitting the same problem,
then it doesn't unlock a new problem for you.