Fix ghost user issue in team invitations
The `find_or_create_by_username` method on User did not handle
usernames in a case insensitive way, even though the database properly
handles case insensitive usernames.

The problem was that the subtraction method was being used on the array
of usernames found in the database against the usernames of the people
being invited.

e.g. If the users 'alice', 'Bob', and 'CHARLIE' were all members, and
someone invited 'alice', 'bob', and 'charlie' to their team, then
we would end up with invitations to all three existing users, plus two
new ghost users, 'bob', and 'charlie'.

When Bob and CHARLIE go to accept the invitation, the code would try to
accept an invitation for CHARLIE, but the the acceptance code does a
(case insensitive) lookup for CHARLIE, and finds the invitation for the
(newer) ghost user, charlie. The ghost user becomes a user on the team,
the site redirects CHARLIE to the team page... which they don't have
access to.

Fixes #2328 and #2467.
Existing ghost users will need to be deleted manually.