Use archive flag in profile and account
The archive flag replaces the state field. Instead of showing tables for current, hibernating,
and completed, we show only tables for current and archived.

The profile and account pages now link to the exercise rather than the submission.

This removes the comment count, since that is a bit ambiguous when we're talking about an
exercise (is this all the comments on all the submissions? just the comments on the latest iteration?).