Replace done/hibernate with archive in iteration page
Archive is easier to understand than 'done' or 'hibernate', and this will allow us to simplify a lot
of other surrounding code.

This changes the "Manage" dropdown, and also cascades some changes to
other parts of the application, so that we don't talk about "archive" in
some places and "completed" or "hibernate" in others.