Write to views table when someone looks at a submission
We are not tracking read/unread at the iteration level, but at the exercise level. This
means that if someone comments on an older submission, this might not be immediately obvious
when you click on the exercise, since that will take you to the most recent iteration.

We will likely redesign the submission/code page, however, to make this more intuitive and useful.