Add database field for last activity description
This is a dirty hack to provide more detail about what the last activity was about when
looking at the inbox.

Without it, we would have to do some nasty queries to get at the data that we need.

Eventually this will be replaced with a proper activity stream (see #1899).