Fix bug preventing teams from being deleted. Fixes #2592
The HTTP form was defined with method="DELETE", which doesn't exist in HTTP.

This implements the method override hack.

The genesis of this bug is that the functionality was originally
implemented asynchronously using JavaScript, and the bug was
introduced when the team management behavior was simplified and
collected into a separate page which just uses regular synchronous
HTTP functionality.