Adding API endpoint for `exercism status $language` CLI command
This adds a very rough, but working, implementation of the endpoint that goes with the new CLI command `exercism status $language`. I wanted to submit this as is now before I start to refactor to make sure this is generally on the right path. I also need to write tests for this endpoint, but since I haven't use MiniTest before, that will have to come a little later.