Remove langauge option from #md helper
Now, the `#md` helper only accepts Markdown. We no longer try to
fence off input with a supplied language. The last remaining place
the argument was used has been converted to `#syntax`.

The markdown helper test has been removed as the functionality
is now tested by the new syntax helper as well as the
Markdown to HTML converter.