Work around nil exception in dashboard
The problem sometimes has a nil language, because of a bug in the API
where people submitting with an old version of the CLI on a windows machine don't
get the track detected correctly. I still need to fix the bug, but at least this
won't blow up as badly on their dashboard.