Add completed exercises in languages page.
This PR adds the feature requested in issue #2780. When a user is logged
in the page shows the completed exercises out of the total for each
language. This label appears only on active exercises, upcoming ones
keep the "# problems" tag. If the user is not logged in all the
languages in the page show the "# problems" tag.