Add task to fix broken iterations
There's an old version of the CLI that the API isn't handling correctly,
and we end up with submissions and user_exercises with an empty language,
and a slug that has the track id instead of the slug.

This repairs those accounts, until we can get the underlying bug fixed.