Update statistics endpoint to scope exercises
The statistics endpoint currently returns user.exercises, which includes exercises with iteration counts of 0.  Because of this, the returned data is incorrect as it lists all fetched exercises as completed.

By scoping into exercises.current, exercises with 0 iterations will be excluded and will mirror the data shown on the user/show view.