Subset of Style changes.
I wanted to get this small percentage of changes up here.  Very little
in the way of actual style changes, but marking where Rubocop indicates
we are needing some improvement.

This puts it in our view, and gets our attention, so when we do work on
the code, we know we can do some touch ups while we are there.

Hiding in the hidden todo file doesn't do that for us.