Update team manage to handle sole manager scenario
This commit tweaks the Manage page of a team to do two things:

- Display the current user in the list of managers if they are one so
  that they can see they are a manager, as well as remove themself as a
  manager if there are other managers.
- If the current user is the only manager, a disabled removal button is
  shown with a tooltip explaining that the user must add another manager
  before they can remove themself as the last manager.

This change was made so that when a manager is trying to leave a team
they know they are a manager and that they must add a new manager before
removing themself as a manager.

This commit also adds an acceptance test for viewing the Manage page of
a team and asserting the current user is in the list, as well as that
there is a disabled remove button since that user is the sole manager.

Fixes https://github.com/exercism/exercism.io/issues/2836