Fix redirect when path is a github callback. Fixes #2842 (#2849)
* Fix redirect when the path is a github callback

This redirect was essentially swallowing the github callback so that
user wouldn't get authenticated when they log in from `/please-login`.

* Add spec for guest redirect when submitting

* Correct order of comment check in local storage

When a user doesn't have a session, we were first checking the value of
the `textarea` and storing it in local storage. This was happening
_before_ we restored the comment _already_ in local storage. By
switching the order of the checks, the comment will be restored as
expected.