Provide 404 for unstarted track in status endpoint
If you don't have any submissions on a language, then we don't have any
recent submissions. If we don't have any recent submissions, then the timestamp
is empty under 'recent', and the CLI blows up.

Maybe we should change how this works and be able to tell you that you
have fetched exercises X, Y, and Z, even if you haven't submitted anything.

If so we need to handle missing recents in the CLI.

See exercism/cli#308