Create test:everything rake task
This adds a new rake task, `rake test:everything`, that runs all of the
test suite checks (MiniTest, RuboCop, and Lineman CI currently). The
goal of this new rake task is to have a single command to run on CI and
locally to ensure that everything is working as expected before a PR is
opened and merged.

Here is a list of detailed changes:

- The Travis CI config was changed to use the new rake task so that
  changes to the rake task do not also have to be updated in the CI
  config.
- Moved the test rake namespace into its own lib/tasks/test.rake file
  since it is getting pretty large.
- Created a `rake test:minitest` task to run the Ruby MiniTest test
  suite and made it use conditionals for Coveralls reporting on CI.
- Created a `rake test:lineman` rake task for running the lineman spec
  check.
- Set the default rake task to `test:everything` so running `bundle exec
  rake` runs every test check.