Rip out fetched tracking
I realized that all the features around tracking/status are
completely haphazard, and need to be thought through from a UX perspective.
We have a number of really annoying bugs that are just not worth
having the features for.

First of all, the dashboard is unusable because we mark every single
language track as having at least one thing fetched. Useless.

Second of all, the HTTP calls back and forth for tracking fetched
status makes the 'exercism fetch' command ridiculously slow
if you've solved a lot of exercises.