Add test:everything rake task (#2871)
* Add RuboCop rake task

This creates a new rake task that handles requiring RuboCop's rake task
so `bundle exec rake rubocop` can be run. The motivation for this
addition is so that other rake tasks can invoke the RuboCop rake task.

* Create test:everything rake task

This adds a new rake task, `rake test:everything`, that runs all of the
test suite checks (MiniTest, RuboCop, and Lineman CI currently). The
goal of this new rake task is to have a single command to run on CI and
locally to ensure that everything is working as expected before a PR is
opened and merged.

Here is a list of detailed changes:

- The Travis CI config was changed to use the new rake task so that
  changes to the rake task do not also have to be updated in the CI
  config.
- Moved the test rake namespace into its own lib/tasks/test.rake file
  since it is getting pretty large.
- Created a `rake test:minitest` task to run the Ruby MiniTest test
  suite and made it use conditionals for Coveralls reporting on CI.
- Created a `rake test:lineman` rake task for running the lineman spec
  check.
- Set the default rake task to `test:everything` so running `bundle exec
  rake` runs every test check.

* Auto-correct RuboCop offenses

The command `bundle exec rubocop -a` was run to auto-correct the fixable
offenses. I ran the auto-correct because the preceding commits add
RuboCop to the CI test suite, so RuboCop needs to pass.

* Manually fix RuboCop offenses

This commit goes through and cleans up RuboCop offenses that were not
auto-correctable.

The changes are:

- remove unnecessary disabling of an offense
- adding an inline disabling of two offenses to the HTML Rouge formatter
- Tweaking HTML Rouge formatter condition to make RuboCop happy
- Making the fuzzy time test use a var for more than 23 hours ago to
  improve the intent and remove offense
- Changing namespace module style for DB config test