Add a link from exercises table to track
I was finding myself going to the exercises table to see which ones I'd
completed, then clicking languages, finding the language I'm working on, then
clicking Available Exercises.

This adds a link to the existing Language name to the track. (Tried to deep link
to the exercises section i.e. `/languages/ruby#exercises`, but it looks like
the anchors aren't working onload?)