Set page titles using Javascript (#2824)
* The website doesn't set page titles so using the browser history is
impossible. They all say 'exercism.io'. This change adds some Javascript
to grab the first H1, H2, or H3 and use that to build the page title
setting it to the text value of the header plus " - exercism.io" -- but
only if the page title isn't already set to something special.

* don't change page title on homepage.