Remove unused older_than access on Submission
This removes the Submission#older_than?, as well as the scope
Submission.older_than, which were removed in
97ca0517e39297020289a3803788ce980f477959 and
c7758501bc5140e8c95a7ea1e5cafb58127a8944, respectively.

Additionally, this removes a RuboCop setting on the same Submission
file, as RuboCop was failing due to unnecessary disabling of
Metrics/ClassLength.