Add /version route that displays meta information
Add a '/version' route that returns a json string with information about
the current build.
(This is similar to what is provided by http://x.exercism.io/)

Example:
```json
{"repository":"https://github.com/exercism/exercism.io","contributing":"https://github.com/exercism/exercism.io/blob/master/CONTRIBUTING.md","build_id":"ef04547"}
```

In order for the app to display the current build an environment
variable needs to be set on the server before running the app.

```bash
export BUILD_ID=$(git rev-parse --short HEAD)
```