Sort /languages/X/contribute problems by implementations
For #2989

This is not a button to change sorting as mentioned in the issue,
but a quick ruby fix which sorts them using number of implementations
by default. I think this looks good by default. If you think otherwise,
I can put together a button and some `js` to sort in place.