Case insensitive solution paths (Fixes #2468)
When using an old version of the CLI that does not submit the track id
as part of the data, Exercism tries to guess the track from the path of
the submitted filename.

However paths could come in with mixed case names and since all language
track_id's are lowercase track_id matching would fail.

This patch makes sure the deduced track_id is all lowercase.

This also required modification to the Iteration class due to
duplication of path parsing logic.