Add per_page to TeamStream and pagination test
We did need to pass per_page=2 in the url to test this, but we also had to change something on the
TeamStream model to make that change take effect.

```ruby
class TeamStream
  attr_reader :team, :viewer_id, :page, :track_id, :user_id, :username, :slug
  attr_accessor :per_page
  def initialize(team, viewer_id, page=1, per_page=50)
    @team = team
    @viewer_id = viewer_id
    @page = page.to_i
    @per_page = per_page.to_i
  end
```
`per_page` is a new parameter. It was previously always set to 50. Now it can be set when you initialize the TeamStream. We also changed the team routes to pass in the `per_page` URL param to the initializer.