Separating out language sections into separate routes
https://github.com/exercism/exercism.io/issues/3035

Problem:

Right now if you go to exercism.io/languages/:track_id there is a menu
in the sidebar. E.g. http://exercism.io/languages/ruby

Each of these is a section in the same page, and we can't link directly
to them. Also, they break the back button in some browsers. If you click
into one, you can't use the back button to get to the previous place.

I think that it would be useful to have actual endpoints and urls for
each of these.

Proposal:

In order to separate out the sections into urls, I decided to create new
partials for each section. The following were created:
  - about
  - contribute
  - exercises
  - help
  - installing
  - learning
  - resources
  - tests

A new route was introduced in order to render the section. This route
will take in the `:track_id` and a new parameter called `:topic`. The
topic will correspond to the section that it will load. A check was also
added to the route to make sure to render the `not_found` page if an
invalid section was added.

This change also removes the anchors from the `languages/languages.rb`
file and now loads the partials that were created