Fix overlapping test names
The tests for team streams, track streams, and the filters for both streams all had conflicting
class names, which means that they were reopening the same class. Since some tests had the same
name, some tests were not being run, which was hiding failures.

This fixes both the test names and the failures.