Flesh out the track stream filters test
This makes it a little bit easier to understand what we're testing, or at least
to remind us about it later.

I changed the viewpoint to always be Alice, since it was confusing that it was sometimes
Alice, sometimes Bob.

I also made room for the watermarks, which I'm working on in a separate branch. That
way the change to the tests will be as small as possible when we do add the watermarks functionality.