Reverse counting logic in stream filters
We count all the exercises that are marked as 'read' so that we can compute the
'unread' count for the sidebar navigation in each category.

Since some things will be marked as read both based on a watermark and
based on direct views, we would double count them. To avoid that we need
to either exclude the relevant views when counting the watermarks, or
exclude the relevant watermarks when counting the relevant views.

The performance of using NOT IN against the views table (7MM rows) was
dismal to the point of taking down production when certain users
accessed the activity stream.

This reverses the logic so we're excluding watermarks when counting
direct views.