Don’t fail if development DB already exists
Previously we were failing if $CHILD_STATUS wasn’t success but
that can be the case simply when the config.database already exists in
psql.

With this we simply check to see if we have a matching database when
the status isn’t success, and output a helpful feedback message.

This allows bin/setup to be run multiple times without failure.