Use the excercism color on CTA buttons
This idea here is the key color has much more contrast and should attract clicks
whereas the current CTA buttons sort of fade into the background. They don't
give enough of a hint to visitors that they should follow them.

This is of course an experiment and I don't mean to change things without a
discussion, please feel free to critique this and reject it if doesn't make
sense.