Reduce the height of the header
Spacing is not a bad thing but this was pushing the list of features,
which is what Exercism is about, way too low on the page (I think).

This retains hefty margins (specified in ems to avoid pixel-focused
design and eventually make this more responsive to viewports) but
trims unnecessary or inconsistent ones.