Center and emphasize feature list
This gives some room to breathe to the feature list, increases the
font size a notch, ditto for the vertical spacing.

Finally it centers (at least on large viewports, some work is needed to make
this look good on smaller viewports but the current styles don’t really
address this much) the list of features in the page while keeping them
individuall left-aligned for legibility and to hug their relevant icon.

I also trimmed the copy for the feature list in the previous commit so that
their horizontal size would be consistent.