Use Trackler for Tracks Route
Helps with https://github.com/exercism/exercism.io/pull/3164

This removes the call to `Xapi` from the route that checks the status of
the language. It now uses the Trackler gem to determine if the the track
exists.

The variable name of the route was also changed from `id` to `language`
so that it makes a little more sense.