Disambiguate "missing track" and "no such topic"
We have a particular set of documentation topics for each track.

If you ask for a topic outside of that set, we were showing you the
page that says "we don't have that track, but Katrina can make you one".

It's not a track, it's a topic within a track.

This changes it to show you the documentation page for that track, with
an error message that makes more sense ("no such topic in the context of
track X").