Consolidate track todos with track documentation
The track todos were linked to from the track documentation page,
however the actual page was full-width, without the breadcrumbs or
the sidebar navigation for the track documentation.

This doesn't fix the scannability of the todos, but it does bring
this into line with the rest of the documentation visually/contextually.