Change so that most recent exercises for a language show at the top.
Change the default sort order to last to first completed.  This makes sure that the most likely information desired is the easiest to find at the top of the list.  This maintains the ability to sort from first to last completed.  This is the second half of issue: exercise in progress sort order #3246