Add task to merge accounts
By default if you change your username on GitHub, then we will update
the username on Exercism the next time you log in.

However, sometimes people create a new GitHub account and either just
leave the old one in place or delete it, and then want to merge their
two exercism accounts.

When someone makes the request we should be very careful that the two
accounts do, indeed, belong to the same person. If the older account no
longer exists, then it's less important to be able to connect this to
the same email address in both cases, but whatever we do, we should
pay attention.

Fixes #3242