Upgrade Ruby to 2.3.3
Previous attempt at upgrading to 2.3.0 was reverted due to lot of
errors in the logs (4b96e0b)

I saw following errors this time:
```
Thread.exclusive is deprecated, use Mutex
```

And found that these were coming out of `sinatra-contrib` which was
fixed in `1.4.7`:
https://github.com/sinatra/sinatra-contrib/issues/183

I suspect the errors were same with `2.3.0` at that time which were
intercepted by newrelic (we no more use newrelic).

Upgrading `sinatra-contrib` to `1.4.7` fixed the errors.