Only store comment in localstorage if it has changed
A bug was identified
(https://github.com/exercism/exercism.io/issues/3289) where a comment
persists in localstorage after submitting. This occurs due to the
keydown event to submit and clear localstorage, firing before the keyup
event that stores the comment in localstorage. This is somewhat
difficult to test locally as you need to insert a `sleep` method into
`app/routes/comments.rb` to ensure the server takes sufficiently long to
respond that `keyup` has time to execute.

To resolve this we only store a comment if it has changed, which is not
the case with a submission.